Füge checkNetBalanceNegative-Test hinzu

+++ src/main/java/dbaccess/DBAccessWarning.java
@@ -38,8 +38,8 @@ public class DBAccessWarning {
 
 	// Methode zur Prüfung, ob die Ausgaben höher sind als die Einnahmen
 	public boolean checkNetBalanceNegative() {
-		TypedQuery<Long> query = entityManager.createNamedQuery("checkNetBalance", Long.class);
-		Long result = query.getSingleResult();
+		TypedQuery<Double> query = entityManager.createNamedQuery("checkNetBalance", Double.class);
+		Double result = query.getSingleResult();
 		return result != null && result > 0;
 	}
 }
